added setting global font color for a message file
The font color can only be set once for a message file, because there's logic to preserve the coloring that already existed in the vanilla file. But since the text editor tool creates a new copy of the file on save anyway, you should always have the unmodified message file available. Just keep up with an unset copy.

Removed a copied file that was unnecessary as well.